Moved SineUI into sine_sys_id_ui.py. Fixed circular imports in sine_syd_id_environments.py
